Remove deprecated UsersLookupByEmailResponse parent class (deprecated in v1.1)
