+++ slack-api-client/src/main/java/com/slack/api/methods/response/users/UsersLookupByEmailResponse.java
@@ -1,4 +1,17 @@
 package com.slack.api.methods.response.users;
 
-public class UsersLookupByEmailResponse extends com.slack.api.methods.response.channels.UsersLookupByEmailResponse {
+import com.slack.api.methods.SlackApiResponse;
+import com.slack.api.model.User;
+import lombok.Data;
+
+@Data
+public class UsersLookupByEmailResponse implements SlackApiResponse {
+
+    private boolean ok;
+    private String warning;
+    private String error;
+    private String needed;
+    private String provided;
+
+    private User user;
 }
